Fix dtype mismatch in MoE EP accumulation path
Cast weighted_output to float32 before index_add_ into the FP32
global_results accumulator. Expert output is BF16 (from INT4 dequant),
matching the FP32-accumulate → BF16-allreduce pattern used across all
other MoE paths in the codebase.

+++ batchgen/models/moonshotai/kimi_k25/assets/modeling_deepseek.py
@@ -2057,7 +2057,7 @@ class DeepseekV3MoE_Decoding_FP8(nn.Module):
 
 			# Weighted accumulation into results
 			weighted_output = expert_output * expert_weights.unsqueeze(-1)
-			global_results.index_add_(0, expert_token_idx, weighted_output)
+			global_results.index_add_(0, expert_token_idx, weighted_output.float())
 
 		global_results = global_results.to(torch.bfloat16)
